refactor: 중복 형변환 제거

+++ src/main/java/com/bidnbuy/server/service/AdminUserService.java
@@ -62,7 +62,7 @@ public class AdminUserService {
         return PagingResponseDto.<AdminUserListDto>builder()
                 .data(dtoList)
                 .totalPages(totalPages)
-                .totalElements((long) totalElements)
+                .totalElements(totalElements)
                 .currentPage(currentPage)
                 .pageSize(pageSize)
                 .isFirst(isFirst)
